ai.audio/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
aiofiles==23.2.1
azure-cognitiveservices-speech==1.34.0
google-cloud-texttospeech==2.16.3
elevenlabs==0.2.26
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
TTS>=0.22.0
torch==2.5.1
torchaudio==2.5.1
